+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 #### Documentation
  - Added documentation for creating and using custom components, including an example PR.
 
+#### New Components
+ - Added ExtraDataBatcher that allows batching items using the extra_data field of the items.
+
 #### Features
  - Added a max_submissions option to schema configs and scheduling to allow slower rollouts for changes that can prevent overwhelming code review.